refactor: created index.ts to export group components and functions
