+++ src/features/register-player/game-styles/helpers/index.ts
