+++ src/features/register-player/game-styles/hooks/index.ts
@@ -1,3 +1,2 @@
-export { useFetchData } from './useFetchData'
 export { useFormGameStyles } from './useFormGameStyles'
 export { useSubmitGameStyles } from './useSubmitGameStyles'
