+++ src/features/register-player/player-registration/helper/index.ts
@@ -4,4 +4,3 @@ export { generateDaysOfAMonth } from './generateDaysOfAMonth'
 export { generateXYearsUntillToday } from './generateXYearsUntillToday'
 export { monthsArray } from './monthsArray'
 export { getSecureLocalStorage, setSecureLocalStorage } from './getLocalStorage'
-export { encode, decode } from './encrypton'
